1.3.0 版本优化、修改Kafka消费者Bean等获取方式

+++ src/main/java/tech/mhuang/ext/kafka/admin/KafkaBuilder.java
@@ -35,7 +35,7 @@ public class KafkaBuilder {
         /**
          * 是否开启当前的生产者 默认不开启
          * @param enable 开启
-         * @return
+         * @return ProducerBuilder
          */
         public ProducerBuilder enable(boolean enable){
             this.producer.setEnable(enable);
@@ -45,7 +45,7 @@ public class KafkaBuilder {
         /**
          * 设置生产者服务、不设置根据通用的进行配置（通用默认是本地127.0.0.1:9200)
          * @param servers 服务ip端口
-         * @return
+         * @return ProducerBuilder
          */
         public ProducerBuilder servers(String servers) {
             this.producer.setServers(servers);
@@ -55,7 +55,7 @@ public class KafkaBuilder {
         /**
          * 设置生产序列化key的类、默认apache string的序列化类
          * @param keySerializer 序列化
-         * @return
+         * @return ProducerBuilder
          */
         public ProducerBuilder keySerializer(String keySerializer) {
             this.producer.setKeySerializer(keySerializer);
@@ -65,7 +65,7 @@ public class KafkaBuilder {
         /**
          * 设置生产序列化value的类、默认apache string的序列化类
          * @param valueSerializer 序列化
-         * @return
+         * @return ProducerBuilder
          */
         public ProducerBuilder valueSerializer(String valueSerializer) {
             this.producer.setValueSerializer(valueSerializer);
@@ -75,7 +75,7 @@ public class KafkaBuilder {
         /**
          * 设置应答、默认all
          * @param acks 应答
-         * @return
+         * @return ProducerBuilder
          */
         public ProducerBuilder acks(String acks) {
             this.producer.setAcks(acks);
@@ -168,7 +168,7 @@ public class KafkaBuilder {
             return this;
         }
 
-        public ConsumerBuilder pull(Integer pull) {
+        public ConsumerBuilder pull(Long pull) {
             this.consumer.setPull(pull);
             return this;
         }
